// src/lib/backend/api/api.ts
export interface V1AdminProductsIdGet200ResponseDataDataStoresInner {
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'updatedAt': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'deletedAt'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressCity'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressCountry'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressState'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressLine1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressLine2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'addressPostalCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'openingHours'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof V1AdminProductsIdGet200ResponseDataDataStoresInner
     */
    'contactNumbers'?: string | null;
}
// ...
export interface V1AdminStoresGet200Response {
    /**
     * 
     * @type {string}
     * @memberof V1AdminStoresGet200Response
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof V1AdminStoresGet200Response
     */
    'message': string | null;
    /**
     * 
     * @type {V1AdminStoresGet200ResponseData}
     * @memberof V1AdminStoresGet200Response
     */
    'data': V1AdminStoresGet200ResponseData;
}
/**
 * 
 * @export
 * @interface V1AdminStoresGet200ResponseData
 */
export interface V1AdminStoresGet200ResponseData {
    /**
     * 
     * @type {Array<V1AdminProductsIdGet200ResponseDataDataStoresInner>}
     * @memberof V1AdminStoresGet200ResponseData
     */
    'data': Array<V1AdminProductsIdGet200ResponseDataDataStoresInner>;
    /**
     * 
     * @type {number}
     * @memberof V1AdminStoresGet200ResponseData
     */
    'count': number;
}